OCT-111: moved GetCategoryTreeRootsQuery
